+++ chatbots-telegram-api/src/main/java/io/micronaut/chatbots/telegram/api/Audio.java
@@ -64,7 +64,7 @@ public class Audio extends AbstractFile {
 
     @Nullable
     @JsonProperty("file_size")
-    private Integer fileSize;
+    private Long fileSize;
 
     /**
      * Thumbnail of the album cover to which the music file belongs.
@@ -145,7 +145,7 @@ public class Audio extends AbstractFile {
      * @return File size
      */
     @Nullable
-    public Integer getFileSize() {
+    public Long getFileSize() {
         return fileSize;
     }
 
@@ -153,7 +153,7 @@ public class Audio extends AbstractFile {
      *
      * @param fileSize File size
      */
-    public void setFileSize(@Nullable Integer fileSize) {
+    public void setFileSize(@Nullable Long fileSize) {
         this.fileSize = fileSize;
     }
